api-gateway/requirements.txt
fastapi
uvicorn
requests
zeep
grpcio
grpcio-tools
protobuf